+++ tiny-firmware/firmware/main.c
@@ -9,6 +9,8 @@
  *
  */
 
+#include <libopencm3/stm32/desig.h>
+
 #include "skywallet.h"
 #include "oled.h"
 #include "bitmaps.h"
@@ -25,7 +27,9 @@
 #include "fastflash.h"
 #include "factory_test.h"
 #include "entropy.h"
+#include "memory.h"
 
+extern uint32_t device_uuid[STM32_UUID_LEN/sizeof(uint32_t)];
 int main(void)
 {
 #if defined(EMULATOR) && EMULATOR == 1
@@ -47,8 +51,11 @@ int main(void)
 	timer_init();
 
 #if !defined(EMULATOR) || EMULATOR == 0
+	desig_get_unique_id(device_uuid);
 	// enable MPU (Memory Protection Unit)
 	mpu_config();
+#else
+    random_buffer((uint8_t *)device_uuid, sizeof(device_uuid));
 #endif  // !defined(EMULATOR) || EMULATOR == 0
 
 #if DEBUG_LINK