+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Get firmware version from `git`, use `tiny-firmware/VERSION` if it's not possible.
+- Read device `uuid` from physical device.
 - Add deploy satge for travis.
 - A makefile target for `release`.
 - A a bash script to get firmware version.